app/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
billiard==3.6.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.6
django-environ==0.8.1
django-silk==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.3
gprof2dot==2022.7.29
gunicorn==20.1.0
humanize==4.2.3
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
